RK3

+++ internal/pkg/order/interfaces.go
@@ -16,7 +16,7 @@ type OrderUsecase interface {
 }
 
 type OrderRepo interface {
-	CreateOrder(context.Context, models.Cart, uuid.UUID, uuid.UUID, int) (models.Order, error)
+	CreateOrder(context.Context, models.Cart, uuid.UUID, uuid.UUID, int64) (models.Order, error)
 	ReadOrderID(context.Context, uuid.UUID) (uuid.UUID, error)
 	ReadOrder(context.Context, uuid.UUID) (models.Order, error)
 	ReadOrdersID(context.Context, uuid.UUID) ([]uuid.UUID, error)
